rename ecom analytics js file to please ublock
